Fix UE 5.6 API compatibility issues
- Replace OnRequestProgress() with OnRequestProgress64() using uint64 parameters
- Update TryGetObjectField() calls to use new const TSharedPtr* signature
- Remove WidgetBlueprint.h include and use FTopLevelAssetPath directly
- Update function signatures in header and implementation files

+++ Plugins/AdastreaDirector/Source/AdastreaDirector/Public/AdastreaLLMClient.h
@@ -143,8 +143,8 @@ private:
 	
 	void OnStreamDataReceived(
 		FHttpRequestPtr Request,
-		int32 BytesSent,
-		int32 BytesReceived,
+		uint64 BytesSent,
+		uint64 BytesReceived,
 		FOnStreamChunk OnStreamChunk
 	);
